backend/requirements.txt
fastapi>=0.68.0,<0.69.0  (X)
uvicorn==0.11.3  (X)
sqlalchemy  (X)
passlib
bcrypt
python-jose
python-multipart
pyclean
pyjwt
typing
pandas==1.4.2 (X)
datetime
python-math
pydantic>=1.8.0,<2.0.0
pymysql
tabulate
janitor
python-jose
alembic
python-dotenv
pyjanitor
gunicorn
httptools
redis  (X)
